// src/model/WeatherData.tsx
interface Root {
  current: Current;
  location: Location;
}

interface Current {
  cloud: number;
  condition: Condition;
  feelslike_c: number;
  feelslike_f: number;
  gust_kph: number;
  gust_mph: number;
  humidity: number;
  is_day: number;
  last_updated: string;
  last_updated_epoch: number;
  precip_in: number;
  precip_mm: number;
  pressure_in: number;
  pressure_mb: number;
  temp_c: number;
  temp_f: number;
  uv: number;
  vis_km: number;
  vis_miles: number;
  wind_degree: number;
  wind_dir: string;
  wind_kph: number;
  wind_mph: number;
}

interface Condition {
  code: number;
  icon: string;
  text: string;
}

interface Location {
  country: string;
  lat: number;
  localtime: string;
  localtime_epoch: number;
  lon: number;
  name: string;
  region: string;
  tz_id: string;
}

interface ErrorResponse {
  error: Error;
}

interface Error {
  code: number;
  message: string;
}

export enum WeatherDegree {
  CELCIUS = 'celcius',
  FARENHEIT = 'farenheit',
}

export type {Root, ErrorResponse};